Add single constant for the name of the datahandle argument
Use only one constant to set the name of the datahandle argument in the
config, since preprocess_setUpCtrlif does not differentiate between the
RHS function and helper functions. Therefore, the name of the datahandle
argument needs to match in the argument list of the RHS function and
helper functions as well as in ctrlif and helper function calls
introduced in an mtree during preprocessing.