Add constants related to the name of the function index
Use two constants, one for the name of the function index argument in
the argument list of helper functions, and another for the name of the
variable which stores the updated function index for a helper function
call. These two constants are independent, since the former is only used
as input for ctrlifs and the function index update function, while the
latter is only used as input for helper function calls.